app/requirements.txt
fastapi
sqlalchemy
uvicorn
passlib
python-jose
openai-whisper
openai
langchain
soundfile
sounddevice
python-dotenv
uvicorn